backend/requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
python-dotenv==1.2.1
python-multipart==0.0.21
requests==2.32.5
httpx==0.28.1
pydantic==2.12.4
openai==2.15.0
PyPDF2==3.0.1
python-docx==1.2.0
svix==1.84.1
lxml==6.0.2
